Lots of whitespace cleanup.
Also fixed an issue where isbndb/book.py was looking for non-existant functions,
specifically it was looking fo IsbndbGetter -- which is an old object that was
factored out at one point when book.py was no longer the main source file
for this module. Also, IsbndbGetter is no longer even a class in this anymore.